+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - fixed Number.MarshalJSON (see #112)'
+- The default type for BLOB is []byte and for CLOB is a string - no need for ClobAsString() option.
 
 ### [2.8.2] - 2018-10-01
 ### Changed